Refactoring: put everything related to robots' current states in proper file.
